Set android:usesCleartextTraffic=true in AndroidManifest.xml. Also add some missing fixes for migration to AndroidX.
Cleartext support is disabled by default so whenever the app sends a request
to http://localhost, it gets blocked by an error that cleartext requests are
not permitted. So we need to set it to true for the app to talk to localhost.

AndroidX-related fixes explained in https://capacitor.ionicframework.com/docs/android/updating#from-1-5-1-to-2-0-0.